Please dont judge me for the negative margin
Im happy with the way this looks, not quite happy with the negative margin. I might give it another smash if I have the time tomorrow morning between the interviews. Also, I did try to install Moti && React-reanimated to give it a quick go on the animations but unfortunately my simulators crash
https://github.com/expo/expo/issues/20661
Might be related to this: I had changed the babel, and ive done this many times before but this seems to not like me today